requirements.txt
# Tek Sigma Kuralı - SigmaHQ Benzerlik Analizi
# Single Sigma Rule - SigmaHQ Similarity Analysis

# Core Dependencies
requests>=2.31.0
pymongo>=4.6.0
PyYAML>=6.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# AI & ML
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Optional: Progress bars
tqdm>=4.66.1

# Optional: Enhanced logging
colorlog>=6.8.0